Introduce backoff time to throttling algorithm
This commit introduces backoff time to the additional throttling algorithm.
When there is an issue with gh api access, an algorithm that was supposed to work
locally (in timespan of around 10-60 seconds per queue) is co-responsible for forming
a massive queue of messages, scheduled to be handled in the future. This is unwanted
behavior. When there is an external problem and the bigger queue forms, the algorithm will be
temporarily disabled. User can choose the value and therefore indirectly steer a length of the queue.

Signed-off-by: Jakub Guzik <jguzik@redhat.com>